Fix canvas resize, move major send-functions into new service 'sender'
