+++ app/services/sender.js
@@ -1,0 +1,1 @@
+export { default } from 'explorviz-frontend-extension-vr/services/sender';
